requirements.txt
httpx
pymilvus
tqdm
python-dotenv
boto3